virtio-mem: introduce VIRTIO_MEM_F_PERSISTENT_SUSPEND
Before, the behavior while suspending to a deep sleep state and waking
up was not specified. For example, in x86 QEMU VM all devices receive a
reset request during wake-up. This would lead to unplugging of all the
plugged memory blocks. Due to this, suspending is disallowed in the
Linux Kernel, when plugged memory is present and
VIRTIO_MEM_F_PERSISTENT_SUSPEND feature flag is not advertised by the
virtio-mem device.

This new flag should signal to the guest driver, that the device can
correctly suspend to a deep sleep state and then wake up without
disrupting the plugged memory blocks. This feature flag is already
supported in the Linux Kernel [1] and should be soon also supported in
QEMU [2].

[1]: https://lore.kernel.org/all/20240318120645.105664-1-david@redhat.com/
[2]: https://lore.kernel.org/all/20240806160756.182524-1-jmarcin@redhat.com/

Signed-off-by: Juraj Marcin <jmarcin@redhat.com>